+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # @3plate/graph
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @3plate/graph-react@0.1.18
+  - @3plate/graph-core@0.1.17
+  - @3plate/graph-angular@0.1.17
+  - @3plate/graph-vue@0.1.17
+
 ## 0.1.17
 
 ### Patch Changes