+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @3plate/graph-vue
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @3plate/graph-core@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes